Add alternating searching indicator to status column. Fix one lock
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@5354 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/ManagedSearch.cpp
@@ -143,15 +143,15 @@ void CManagedSearch::Start()
 
 void CManagedSearch::Stop()
 {
+	CSingleLock pLock( &SearchManager.m_pSection );
+	if ( !pLock.Lock( 100 ) ) return;
+
+	SearchManager.Remove( this );
 	if ( m_bActive )
 	{
 		m_bActive = FALSE;
 	    Datagrams.PurgeToken( this );
 	}
-	
-	CSingleLock pLock( &SearchManager.m_pSection );
-	pLock.Lock( 1000 );
-	SearchManager.Remove( this );
 }
 
 //////////////////////////////////////////////////////////////////////
